// src/api/dto/match-timeline-dto.ts
type MatchTimelineDto = {
  metadata: MetadataDto;
  info: InfoDto;
};

type InfoDto = {
  frameInterval: number;
  frames: FrameDto[];
};

type FrameDto = {
  events: EventDto[];
  participantFrames: ParticipantFramesDto;
  timestamp: number;
};
// ...
type PositionDto = {
  x: number;
  y: number;
};
// ...
export type ChampionKillEventDto = {
  assistingParticipantIds: number[];
  bounty: number;
  killStreakLength: number;
  killerId: number;
  position: PositionDto;
  shutdownBounty: number;
  timestamp: number;
  type: "CHAMPION_KILL";
  victimDamageDealt: DamageDto[];
  victimDamageReceived: DamageDto[];
  victimId: number;
};

export type DamageDto = {
  basic: boolean;
  magicDamage: number;
  name: string;
  participantId: number;
  physicalDamage: number;
  spellName: string;
  spellSlot: number
  trueDamage: number;
  type: string;
};